+++ src/test/kotlin/com/cmgapps/gradle/reporter/JsonReportShould.kt
@@ -6,15 +6,34 @@
 
 package com.cmgapps.gradle.reporter
 
+import com.cmgapps.gradle.JSON_REPORT_NAME
+import com.cmgapps.gradle.extension.ReportTask
+import com.cmgapps.gradle.extension.ReportTaskExtension
+import com.cmgapps.gradle.model.Package
+import org.gradle.api.Task
+import org.gradle.api.file.ProjectLayout
+import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.provider.Property
+import org.gradle.kotlin.dsl.property
+import org.gradle.testfixtures.ProjectBuilder
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.`is`
 import org.intellij.lang.annotations.Language
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 
-class JsonReporterShould : OutputStreamTest() {
+@ExtendWith(ReportTaskExtension::class)
+class JsonReportShould : OutputStreamTest() {
+    @ReportTask
+    lateinit var task: Task
+
     @Test
     fun `report outdated and latest`() {
-        JsonReporter(outdated = outdated, latest = latest).writePackages(outputStream)
+        TestJsonReporter(
+            task = task,
+            outdated = outdatedPackages,
+            latest = latestPackages,
+        ).writePackages(outputStream)
 
         @Language("json")
         val expected =
@@ -42,7 +61,11 @@ class JsonReporterShould : OutputStreamTest() {
 
     @Test
     fun `report outdated only`() {
-        JsonReporter(outdated = outdated, latest = emptyList()).writePackages(outputStream)
+        TestJsonReporter(
+            task = task,
+            outdated = outdatedPackages,
+            latest = emptyList(),
+        ).writePackages(outputStream)
 
         @Language("json")
         val expected =
@@ -65,7 +88,11 @@ class JsonReporterShould : OutputStreamTest() {
 
     @Test
     fun `report latest only`() {
-        JsonReporter(outdated = emptyList(), latest = latest).writePackages(outputStream)
+        TestJsonReporter(
+            task = task,
+            outdated = emptyList(),
+            latest = latestPackages,
+        ).writePackages(outputStream)
 
         @Language("json")
         val expected =
@@ -84,4 +111,26 @@ class JsonReporterShould : OutputStreamTest() {
             `is`(expected),
         )
     }
+}
+
+private class TestJsonReporter(
+    task: Task,
+    override var outdated: List<Package>,
+    override var latest: List<Package>,
+) : JsonReport(JSON_REPORT_NAME, task) {
+    override fun getRequired(): Property<Boolean> =
+        ProjectBuilder
+            .builder()
+            .build()
+            .objects
+            .property()
+
+    override fun getOutputLocation(): RegularFileProperty =
+        ProjectBuilder
+            .builder()
+            .build()
+            .objects
+            .fileProperty()
+
+    override fun getProjectLayout(): ProjectLayout = ProjectBuilder.builder().build().layout
 }