+++ src/test/kotlin/com/cmgapps/gradle/reporter/TextReportShould.kt
@@ -6,15 +6,33 @@
 
 package com.cmgapps.gradle.reporter
 
+import com.cmgapps.gradle.PLAIN_TEXT_REPORT_NAME
+import com.cmgapps.gradle.extension.ReportTask
+import com.cmgapps.gradle.extension.ReportTaskExtension
 import com.cmgapps.gradle.model.Package
+import org.gradle.api.Task
+import org.gradle.api.file.ProjectLayout
+import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.provider.Property
+import org.gradle.kotlin.dsl.property
+import org.gradle.testfixtures.ProjectBuilder
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.`is`
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 
-class TextReporterShould : OutputStreamTest() {
+@ExtendWith(ReportTaskExtension::class)
+class TextReportShould : OutputStreamTest() {
+    @ReportTask
+    lateinit var task: Task
+
     @Test
     fun `should print header`() {
-        TextReporter(emptyList(), emptyList()).writePackages(outputStream)
+        TestTextReport(
+            task = task,
+            emptyList(),
+            emptyList(),
+        ).writePackages(outputStream)
 
         assertThat(
             outputStream.asString(),
@@ -28,9 +46,10 @@ class TextReporterShould : OutputStreamTest() {
 
     @Test
     fun `should print outdated and latest`() {
-        TextReporter(
-            outdated = outdated,
-            latest = latest,
+        TestTextReport(
+            task = task,
+            outdated = outdatedPackages,
+            latest = latestPackages,
         ).writePackages(outputStream)
 
         assertThat(
@@ -49,8 +68,9 @@ class TextReporterShould : OutputStreamTest() {
 
     @Test
     fun `should print outdated only`() {
-        TextReporter(
-            outdated = outdated,
+        TestTextReport(
+            task = task,
+            outdated = outdatedPackages,
             emptyList(),
         ).writePackages(outputStream)
 
@@ -68,16 +88,10 @@ class TextReporterShould : OutputStreamTest() {
 
     @Test
     fun `should only print latest`() {
-        TextReporter(
-            outdated =
-                emptyList(),
-            listOf(
-                Package(
-                    name = "latest list",
-                    currentVersion = "1.0.0",
-                    availableVersion = "1.0.0",
-                ),
-            ),
+        TestTextReport(
+            task = task,
+            outdated = emptyList(),
+            latest = latestPackages,
         ).writePackages(outputStream)
 
         assertThat(
@@ -91,4 +105,26 @@ class TextReporterShould : OutputStreamTest() {
             ),
         )
     }
+}
+
+class TestTextReport(
+    task: Task,
+    override var outdated: List<Package>,
+    override var latest: List<Package>,
+) : TextReport(PLAIN_TEXT_REPORT_NAME, task) {
+    override fun getRequired(): Property<Boolean> =
+        ProjectBuilder
+            .builder()
+            .build()
+            .objects
+            .property()
+
+    override fun getOutputLocation(): RegularFileProperty =
+        ProjectBuilder
+            .builder()
+            .build()
+            .objects
+            .fileProperty()
+
+    override fun getProjectLayout(): ProjectLayout = ProjectBuilder.builder().build().layout
 }