+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Added
 
 ### Changed
+- Move from Reporter to [Report API](https://docs.gradle.org/current/javadoc/org/gradle/api/reporting/Report.html)
 
 ### Deprecated
 
 ### Removed
+- All Reporter
 
 ### Fixed